start_point/requirements.txt
pulp==2.7.0
ujson
PyPDF2
PyCryptodome
nltk
tqdm
requests
numpy
scikit-learn
bs4
python-magic
